refactor: response dto 위치 수정
