+++ src/test/java/com/example/sns/post/application/PostQueryServiceTest.java
@@ -4,7 +4,7 @@ import com.example.sns.common.support.ServiceTest;
 import com.example.sns.imagestore.infrastructure.ImageStore;
 import com.example.sns.member.domain.Member;
 import com.example.sns.member.exception.MemberNotFoundException;
-import com.example.sns.post.application.dto.PostResponse;
+import com.example.sns.post.presentiation.dto.PostResponse;
 import com.example.sns.post.domain.CommentRepository;
 import com.example.sns.post.domain.Post;
 import com.example.sns.post.domain.PostRepository;
