+++ src/test/java/com/example/sns/post/presentiation/CommentControllerTest.java
@@ -6,19 +6,25 @@ import com.example.sns.post.exception.CommentNotFoundException;
 import com.example.sns.post.exception.EmptyCommentException;
 import com.example.sns.post.exception.NotCommentAuthorException;
 import com.example.sns.post.exception.PostNotFoundException;
+import com.example.sns.post.presentiation.dto.CommentsResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 
+import java.util.List;
+
 import static com.example.sns.common.fixtures.AuthFixture.ACCESS_TOKEN;
 import static com.example.sns.common.fixtures.CommentFixture.getBasicCommentRequest;
 import static com.example.sns.common.fixtures.CommentFixture.getEmptyContentCommentRequest;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.doThrow;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 class CommentControllerTest extends MockControllerTest {
@@ -111,5 +117,20 @@ class CommentControllerTest extends MockControllerTest {
         mockMvc.perform(delete("/posts/1/comments/1")
                         .header(HttpHeaders.AUTHORIZATION, "Bearer " + ACCESS_TOKEN))
                 .andExpect(status().isBadRequest());
+    }
+
+    @Test
+    @DisplayName("게시글 댓글을 조회할 경우 올바른 데이터와 200을 응답해야 함")
+    void findCommentsTest() throws Exception {
+        //given
+        CommentsResponse response = new CommentsResponse(List.of(), true, 0);
+        given(commentQueryService.findComments(any(), any()))
+                .willReturn(response);
+
+        //when then
+        mockMvc.perform(get("/posts/1/comments")
+                        .header(HttpHeaders.AUTHORIZATION, "Bearer " + ACCESS_TOKEN))
+                .andExpect(status().isOk())
+                .andExpect(content().json(mapper.writeValueAsString(response)));
     }
 }
